+++ packages/fuels-programs/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod calls;
 pub mod contract;
+pub mod executable;
 pub mod responses;
